Refactored Exercise Journal page
